Split activation and deactivation routines into separate classes.
