+++ src/Plugin.php
@@ -38,6 +38,8 @@ class Plugin extends AbstractPlugin implements Composable {
 		// Register hook providers.
 		$this
 			->register_hooks( new I18n() )
+			->register_hooks( new Provider\Activation() )
+			->register_hooks( new Provider\Deactivation() )
 			->register_hooks( new Provider\RewriteRules() )
 			->register_hooks( new Provider\CustomVendor() )
 			->register_hooks( new Authentication\Basic\Request() )
